Feature and a bugfix
Fixes #43 FEATURE: Added a checkbox for hiding contents of the "empty" trays.
Fixes #44 BUG: Looses allocations when pulling out a drive, probably caused because it still existed in the "location" table. Now it gets deleted. Also checks for an older reinserted device and modifies the status in the "disks" table.