Bugfixes
Fixed so that default values are displayed in column and sort if the files has been reset.
Added a check if the file exists before trying to delete it.
Added a touch with new atime to CSS files in use after migration from SQLite to JSON, also added it after restoring old database. This should keep CSS up to date.
Also delete the SQLite database when conversion has succeeded, the file is anyway backed up. Otherwise the file will be converted again if "reset all/wipe" is initiated.